support .npz & .nrsz

+++ examples/save.py
@@ -1,8 +1,8 @@
 import numpy as np
 
 # 1️⃣ 创建一些数组
-arr1 = np.arange(10)            # 0~9 的一维数组
-arr2 = np.ones((3, 3))          # 3x3 全 1
+arr1 = np.arange(10, dtype=np.int32)            # 0~9 的一维数组
+arr2 = np.ones((3, 3), dtype=np.int32)            # 3x3 全 1
 arr3 = np.random.rand(2, 4)     # 2x4 随机浮点数数组
 
 # 2️⃣ 保存为 .npz 文件
